--- libraries/OneWire/Readme.MD
@@ -1,6 +1,0 @@
-# OneWire Library for XMC
-
-OneWire library lets you access 1-wire devices.
-
-## Additional Information
-See also our [documentation](https://xmc-arduino.readthedocs.io/en/latest/builtin-libraries.html).
