+++ libraries/OneWire/examples/DS18x20_Temperature/DS18x20_Temperature.ino
@@ -5,7 +5,7 @@
 // http://www.pjrc.com/teensy/td_libs_OneWire.html
 //
 // The DallasTemperature library can do all this work for you!
-// http://milesburton.com/Dallas_Temperature_Control_Library
+// https://github.com/milesburton/Arduino-Temperature-Control-Library
 
 OneWire  ds(10);  // on pin 10 (a 4.7K resistor is necessary)
 
@@ -17,7 +17,7 @@ void loop(void) {
   byte i;
   byte present = 0;
   byte type_s;
-  byte data[12];
+  byte data[9];
   byte addr[8];
   float celsius, fahrenheit;
