+++ libraries/OneWire/library.properties
@@ -1,10 +1,10 @@
 name=OneWire
-version=2.3.4
+version=2.3.8
 author=Jim Studt, Tom Pollard, Robin James, Glenn Trewitt, Jason Dangel, Guillermo Lovato, Paul Stoffregen, Scott Roberts, Bertrik Sikken, Mark Tillotson, Ken Butcher, Roger Clark, Love Nystrom
 maintainer=Paul Stoffregen
 sentence=Access 1-wire temperature sensors, memory and other chips.
-paragraph=Add Dallas one wire interface for single wire peripherals like DS18B20 (temperature), memory and other devices
+paragraph=
 category=Communication
 url=http://www.pjrc.com/teensy/td_libs_OneWire.html
-architectures=xmc
+architectures=*
